Tad + Step
Separação em tad + modelo step
