+++ MiniMips8BitsMonociclo/multiplexadores.c
@@ -1,0 +1,7 @@
+#include "memoria.h"
+#include "minimips.h"
+#include "step.h"
+#include <string.h>
+#include <stdlib.h>
+#include <math.h>
+#include <stdio.h>
