+++ MiniMips8BitsMonociclo/step.h
@@ -1,0 +1,1 @@
+void step (int *parada, int *pc, struct memoria_instrucao *memInst, BRegs *bancoReg, CTRL *controle, descPilha *pilha, struct estatistica * stat, int * estadoControle);
